+++ config/presentable.php
@@ -1,0 +1,11 @@
+<?php
+
+return [
+
+    /**
+     * Specify the presentable key that will be used when
+     * casting a model to an array.
+     */
+    'key' => 'presentable',
+
+];