+++ src/IsPresentableServiceProvider.php
@@ -10,10 +10,15 @@ class IsPresentableServiceProvider extends ServiceProvider
 {
     public function register(): void
     {
+        $this->mergeConfigFrom(
+            __DIR__.'/../config/presentable.php', 'presentable'
+        );
     }
 
     public function boot(): void
     {
-        //
+        $this->publishes([
+            __DIR__.'/../config/presentable.php' => config_path('presentable.php'),
+        ]);
     }
 }