+++ tests/TestCase.php
@@ -5,7 +5,14 @@ declare(strict_types=1);
 namespace TPG\Tests;
 
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
+use TPG\IsPresentable\IsPresentableServiceProvider;
 
 class TestCase extends OrchestraTestCase
 {
+    protected function getPackageProviders($app): array
+    {
+        return [
+            IsPresentableServiceProvider::class,
+        ];
+    }
 }